Create separate files for quakeml classes and functions
